+++ mpire/progress_bar.py
@@ -165,7 +165,7 @@ class ProgressBarHandler:
                     if self.worker_comms.exception_thrown():
                         _, traceback_str = self.worker_comms.get_exception()
                         self._send_dashboard_update(progress_bar, failed=True, traceback_str=traceback_str)
-                        self.worker_comms.task_done_exception()
+                        self.worker_comms.task_done_exception(progress_bar=True)
                     elif self.worker_comms.kill_signal_received():
                         self._send_dashboard_update(progress_bar, failed=True, traceback_str='Kill signal received')